+++ src/components/daySlots/DaySlots.tsx
@@ -4,6 +4,7 @@ import {
   ReactNode,
   useContext,
   useRef,
+  useState,
 } from "react";
 import { defaultWeekStartsOn } from "../../constants/defaults";
 import { bindWeekDayToNumber } from "../../constants/weekdays";
@@ -55,6 +56,11 @@ function DaySlots(props: TDaySlots) {
     inSelectedRangeParentStyles,
     inSelectedRangeParentClassName,
 
+    inHoveredRangeStyles,
+    inHoveredRangeClassName,
+    inHoveredRangeParentStyles,
+    inHoveredRangeParentClassName,
+
     startOfRangeStyles,
     startOfRangeClassName,
     startOfRangeParentStyles,
@@ -76,6 +82,8 @@ function DaySlots(props: TDaySlots) {
     firstDayOfMonth,
   } = useContext(PickerContext);
 
+  const [hoveredItem, setHoveredItem] = useState<Date | undefined>(undefined);
+
   const {
     locale,
     showOtherDays,
@@ -86,6 +94,7 @@ function DaySlots(props: TDaySlots) {
     weekends,
     weekendSelectable = true,
     weekStartsOn = defaultWeekStartsOn,
+    allowBackwardRange,
   } = config || {};
 
   const gridRef = useRef<HTMLDivElement>(null);
@@ -106,6 +115,14 @@ function DaySlots(props: TDaySlots) {
     }).format(day);
   };
 
+  const IsOtherMonth = (date: Date) => {
+    return !(
+      monthInTheCalendar &&
+      calendar &&
+      IsSameMonth(date, monthInTheCalendar, calendar)
+    );
+  };
+
   const IsSelected = (date: Date) => {
     if (Array.isArray(selectedDay)) {
       return (
@@ -117,13 +134,84 @@ function DaySlots(props: TDaySlots) {
   };
 
   const IsRangeSelected = (date: Date) => {
-    if (Array.isArray(selectedDay)) {
+    if (Array.isArray(selectedDay) && selectedDay?.[0] && selectedDay?.[1]) {
       return isWithinInterval(date, {
         start: selectedDay[0].setHours(0, 0, 0, 0),
         end: selectedDay[selectedDay?.length - 1].setHours(0, 0, 0, 0),
       });
     }
     return false;
+  };
+
+  const IsStartOfRange = (date: Date) => {
+    return Array.isArray(selectedDay) && isSameDay(date, selectedDay?.[0]);
+  };
+
+  const IsEndOfRange = (date: Date) => {
+    return (
+      Array.isArray(selectedDay) &&
+      selectedDay.length === 2 &&
+      isSameDay(date, selectedDay?.[selectedDay?.length - 1])
+    );
+  };
+
+  const IsDisabled = (date: Date, isOtherMonth: boolean) => {
+    return (
+      (minDate && date < minDate) || (maxDate && date > maxDate) || isOtherMonth
+    );
+  };
+
+  /**
+   * (otherDays and weekends) can be Enabled but NOT Selectable
+   */
+  const IsSelectable = ({
+    isInWeekend,
+    isOtherMonth,
+    isDisabled,
+  }: {
+    isInWeekend: boolean;
+    isOtherMonth: boolean;
+    isDisabled: boolean;
+  }): boolean => {
+    if (isOtherMonth) {
+      if (isInWeekend) {
+        return weekendSelectable;
+      }
+      return otherDaysSelectable || false;
+    }
+    if (isInWeekend) {
+      return weekendSelectable;
+    }
+    return !isDisabled;
+  };
+
+  const IsInWeekend = (date: Date) => {
+    return !!weekends
+      ?.map((w) => bindWeekDayToNumber[w])
+      .includes(date.getDay() as Day);
+  };
+
+  const IsFirstDayOfMonth = (date: Date) => {
+    return firstDayOfMonth && isSameDay(date, firstDayOfMonth);
+  };
+
+  const IsInHoveredRange = (date: Date) => {
+    if (!hoveredItem) return false;
+
+    if (
+      !Array.isArray(selectedDay) ||
+      (Array.isArray(selectedDay) && selectedDay?.[0] && selectedDay?.[1])
+    ) {
+      return false;
+    }
+
+    // check if hovered date is before the selected date
+    if (hoveredItem < selectedDay?.[0] && !allowBackwardRange) return false;
+
+    return isWithinInterval(date, {
+      start: selectedDay[0].setHours(0, 0, 0, 0),
+      end: hoveredItem.setHours(0, 0, 0, 0),
+    });
   };
 
   const onClickSlot = (date: Date) => {
@@ -191,6 +279,23 @@ function DaySlots(props: TDaySlots) {
     daysOfMonth?.[0]?.getDay() &&
     (daysOfMonth?.[0]?.getDay() - bindWeekDayToNumber[weekStartsOn] + 7) % 7;
 
+  const handleMouseEnter = (date: Date) => {
+    if (
+      !Array.isArray(selectedDay) ||
+      (Array.isArray(selectedDay) && selectedDay?.[0] && selectedDay?.[1])
+    ) {
+      return;
+    }
+
+    setHoveredItem(date);
+  };
+
+  const handleMouseLeave = () => {
+    if (!hoveredItem) return;
+
+    setHoveredItem(undefined);
+  };
+
   return (
     <div
       className={classJoin(
@@ -209,50 +314,25 @@ function DaySlots(props: TDaySlots) {
 
       {daysOfMonth?.map((date) => {
         const IsToday = isToday(date);
-
-        const isOtherMonth = !(
-          monthInTheCalendar &&
-          calendar &&
-          IsSameMonth(date, monthInTheCalendar, calendar)
-        );
-
+        const isOtherMonth = IsOtherMonth(date);
         const isSelected = !!IsSelected(date);
-
         const isInSelectedRange = IsRangeSelected(date);
-
-        const isStartOfRange =
-          Array.isArray(selectedDay) && isSameDay(date, selectedDay?.[0]);
-
-        const isEndOfRange =
-          Array.isArray(selectedDay) &&
-          selectedDay.length === 2 &&
-          isSameDay(date, selectedDay?.[selectedDay?.length - 1]);
-
-        const isDisabled =
-          (minDate && date < minDate) ||
-          (maxDate && date > maxDate) ||
-          isOtherMonth;
-
-        const isInWeekend = !!weekends
-          ?.map((w) => bindWeekDayToNumber[w])
-          .includes(date.getDay() as Day);
-
-        /**
-         * (otherDays and weekends) can be Enabled but NOT Selectable
-         */
-        const isSelectable = !!(isInWeekend
-          ? weekendSelectable
-          : isOtherMonth
-          ? otherDaysSelectable
-          : !isDisabled);
-
-        const isFirstDayOfMonth =
-          firstDayOfMonth && isSameDay(date, firstDayOfMonth);
+        const isInHoveredRange = IsInHoveredRange(date);
+        const isStartOfRange = IsStartOfRange(date);
+        const isEndOfRange = IsEndOfRange(date);
+        const isDisabled = IsDisabled(date, isOtherMonth);
+        const isInWeekend = IsInWeekend(date);
+        const isFirstDayOfMonth = IsFirstDayOfMonth(date);
+        const isSelectable = IsSelectable({
+          isInWeekend,
+          isOtherMonth,
+          isDisabled,
+        });
 
         const formattedDay = dayFormatter(date);
         const ariaFormattedDay = ariaDayFormatter(date);
 
-        if (!showOtherDays && isOtherMonth) return <div />;
+        if (!showOtherDays && isOtherMonth) return <div key={date.getTime()} />;
 
         /* -------------------------------------------------------------------------- */
         /*                     render custom component if provided                    */
@@ -265,6 +345,7 @@ function DaySlots(props: TDaySlots) {
             isSelectable,
             isDisabled,
             isInSelectedRange,
+            isInHoveredRange,
             isStartOfRange,
             isEndOfRange,
             isInWeekend,
@@ -280,6 +361,7 @@ function DaySlots(props: TDaySlots) {
           ...(isSelectable && selectableParentStyles),
           ...(isSelected && selectedParentStyles),
           ...(isInSelectedRange && inSelectedRangeParentStyles),
+          ...(isInHoveredRange && inHoveredRangeParentStyles),
           ...(isStartOfRange && startOfRangeParentStyles),
           ...(isEndOfRange && endOfRangeParentStyles),
           ...(isInWeekend && weekendParentStyles),
@@ -292,6 +374,7 @@ function DaySlots(props: TDaySlots) {
           ...(isSelectable && selectableStyles),
           ...(isSelected && selectedStyles),
           ...(isInSelectedRange && inSelectedRangeStyles),
+          ...(isInHoveredRange && inHoveredRangeStyles),
           ...(isStartOfRange && startOfRangeStyles),
           ...(isEndOfRange && endOfRangeStyles),
           ...(isInWeekend && weekendStyles),
@@ -303,11 +386,12 @@ function DaySlots(props: TDaySlots) {
           slotParentClassName,
           IsToday && "rhmdp-text-blue-600",
           IsToday && todayParentClassName,
-          isSelectable ? "rhmdp-cursor-pointer" : "rhmdp-cursor-default",
           isSelectable && selectableParentClassName,
           isSelected && selectedParentClassName,
           isInSelectedRange && "rhmdp-bg-[#EAEAEC]",
           isInSelectedRange && inSelectedRangeParentClassName,
+          isInHoveredRange && "rhmdp-bg-[#EAEAEC]",
+          isInHoveredRange && inHoveredRangeParentClassName,
           isStartOfRange && "rhmdp-rounded-s-lg",
           isStartOfRange && startOfRangeParentClassName,
           isEndOfRange && "rhmdp-rounded-e-lg",
@@ -322,12 +406,14 @@ function DaySlots(props: TDaySlots) {
           "rhmdp-p-2 rhmdp-rounded-lg rhmdp-h-full",
           slotClassName,
           IsToday && todayClassName,
+          isSelectable ? "rhmdp-cursor-pointer" : "rhmdp-cursor-not-allowed",
           isSelectable && !isSelected && "hover:rhmdp-bg-gray-300",
           isSelectable && selectableClassName,
           isSelected &&
             "rhmdp-bg-blue-500 hover:rhmdp-bg-blue-500 rhmdp-text-white",
           isSelected && selectedClassName,
           isInSelectedRange && inSelectedRangeClassName,
+          isInHoveredRange && inHoveredRangeClassName,
           isStartOfRange && startOfRangeClassName,
           isEndOfRange && endOfRangeClassName,
           isInWeekend && weekendClassName,
@@ -336,9 +422,11 @@ function DaySlots(props: TDaySlots) {
 
         return (
           <div
-            key={date.toString()}
+            key={date.getTime()}
             className={parentClassNames}
             style={parentStyles}
+            onMouseEnter={() => handleMouseEnter(date)}
+            onMouseLeave={handleMouseLeave}
           >
             <div
               role="button"